+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/listeners/ActiveHandlerManager.kt
@@ -1,7 +1,9 @@
 package com.github.continuedev.continueintellijextension.listeners
 
 import com.github.continuedev.continueintellijextension.nextEdit.NextEditService
+import com.github.continuedev.continueintellijextension.nextEdit.NextEditStatusService
 import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.LogicalPosition
 import com.intellij.openapi.editor.event.CaretEvent
@@ -100,7 +102,7 @@ class ActiveHandlerManager(private val project: Project) : SelectionListener, Ca
 
     // SelectionListener implementation
     override fun selectionChanged(event: SelectionEvent) {
-        if (isHandlingEvent || event.editor.isDisposed) return
+        if (isHandlingEvent || event.editor.isDisposed || !isNextEditEnabled()) return
 
         coroutineScope.launch {
             handleCursorMovement(event.editor, event.newRange.startOffset)
@@ -109,7 +111,7 @@ class ActiveHandlerManager(private val project: Project) : SelectionListener, Ca
 
     // CaretListener implementation
     override fun caretPositionChanged(event: CaretEvent) {
-        if (isHandlingEvent || event.editor.isDisposed) return
+        if (isHandlingEvent || event.editor.isDisposed || !isNextEditEnabled()) return
 
         coroutineScope.launch {
             handleCursorMovement(event.editor, event.caret?.offset ?: return@launch)
@@ -167,6 +169,9 @@ class ActiveHandlerManager(private val project: Project) : SelectionListener, Ca
             println("ActiveHandlerManager: Error handling deliberate cursor movement: ${e.message}")
         }
     }
+
+    private fun isNextEditEnabled() =
+        project.service<NextEditStatusService>().isNextEditEnabled()
 
     fun dispose() {
         clearActiveHandler()
